# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gerenciador de Tarefas API",
        "description": "Documentação da API do Gerenciador de Tarefas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/projects/{projectId}/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Listar tarefas de um projeto",
                "operationId": "af78552a9f35ad5da7b75323f49cf431",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "UUID do projeto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tarefas do projeto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Criar tarefa em um projeto",
                "operationId": "6999c29f908d4e41c59a32593a564d4b",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "UUID do projeto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "assignee_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nova task"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Detalhes da tarefa"
                                    },
                                    "assignee_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "20747cde-02a3-4c9c-af0d-52b6d479e6a2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarefa criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Detalhar tarefa",
                "operationId": "eed58132287c5d424986e08508dfd008",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarefa encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "put": {
                "tags": [
                    "Task"
                ],
                "summary": "Atualizar tarefa",
                "operationId": "d44e1e2baf02cf878103612f21948d41",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Título atualizado"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "DONE"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarefa atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Deletar tarefa",
                "operationId": "cfd6adf8691b62edc1c620192f929b98",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tarefa deletada"
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/api/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Listar todas as tarefas",
                "operationId": "571c0b06a45b78df8bb7e8339773046d",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista global de tarefas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Task": {
                "required": [
                    "id",
                    "project_id",
                    "name",
                    "status",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "78b8e7ad-f8e4-4e2d-a4c1-789611d0b0a7"
                    },
                    "project_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9a2b65a4-ffb2-4e5f-aef6-7c508c9b21d1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Implementar login"
                    },
                    "description": {
                        "type": "string",
                        "example": "Criar tela e endpoint de login"
                    },
                    "status": {
                        "type": "string",
                        "example": "TODO"
                    },
                    "assignee_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "20747cde-02a3-4c9c-af0d-52b6d479e6a2"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-28T22:17:53Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-28T22:17:53Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}